+++ CHANGELOG.md
@@ -6,6 +6,13 @@ All notable changes to this project will be documented in this file.
 
 - [#214](https://github.com/os2display/display-api-service/pull/214)
   - Updated endSessionUrl to be nullable.
+
+- [#213](https://github.com/os2display/display-api-service/pull/213)
+  - Add regionsAndPlaylists and groups to `ScreenInput.php`
+  - Add adding playlist/regions in `ScreenProcessor.php` (as an alternative to sending multiple requests)
+  - Add a function that deletes all playlists for a screen in a specific region (`PlaylistScreenRegionRepository.php`)
+  - Save groups in `ScreenProcessor.php`
+
 - [#193](https://github.com/os2display/display-api-service/pull/193)
   - Adds support for interactive slides.
   - Adds interactivity for creating quick bookings from a slide through Microsoft Graph.